[ feat ] sharedPage 서치바 로직 완성
